services/requirements.txt
nameko==2.4.3
python-decouple==3.0
